SUP003-Intro_to_DevSecOps_with_JFrog_Xray/provisioning/python-demo/src/requirements.txt
Flask==1.1.2
requests<=2.3.0
pyyaml<=5.4.1
